Add ping command

+++ proxy-implementation/src/main/java/com/github/phantompowered/proxy/protocol/play/shared/PacketPlayKeepAlive.java
@@ -56,7 +56,7 @@ public class PacketPlayKeepAlive implements Packet {
         protoBuf.writeVarInt(this.randomId);
     }
 
-    public long getRandomId() {
+    public int getRandomId() {
         return this.randomId;
     }
